requirements.txt
pandas>=2.0
numpy>=1.24
tqdm>=4.66
requests>=2.31
pypdf>=4.0
python-dateutil>=2.9
networkx>=3.2

# 可选：更好的实体对齐（装不上也能用fallback）
sentence-transformers>=2.6
scikit-learn>=1.3

# 可选：训练GNN
torch>=2.2
torch-geometric>=2.5
